refactor: openapi definition are now in a specific folder
