+++ lib/arke_server.ex
@@ -29,7 +29,7 @@ defmodule ArkeServer do
              Do NOT define functions inside the quoted expressions
              below. Instead, define any helper function in modules
              and import those modules here.
-             """ && false
+             """
 
   def controller do
     quote do
