+++ lib/arke_server/controllers/arke_controller.ex
@@ -15,9 +15,13 @@
 defmodule ArkeServer.ArkeController do
   @moduledoc """
              Documentation for  `ArkeServer.ArkeController`.
-             """ && false
+             """
 
   use ArkeServer, :controller
+
+  # Openapi request definition
+  use ArkeServer.Openapi.Spec, module: ArkeServer.Openapi.ArkeControllerSpec
+
   alias Arke.{QueryManager, LinkManager, StructManager}
   alias Arke.Boundary.ArkeManager
   alias UnitSerializer
@@ -28,136 +32,14 @@ defmodule ArkeServer.ArkeController do
 
   alias OpenApiSpex.{Operation, Reference}
 
-  # ------- start OPENAPI spec -------
-  def open_api_operation(action) do
-    operation = String.to_existing_atom("#{action}_operation")
-    apply(__MODULE__, operation, [])
-  end
-
-  def get_unit_operation() do
-    %Operation{
-      tags: ["Unit"],
-      summary: "Get unit",
-      description: "Get all the units of an Arke",
-      operationId: "ArkeServer.ArkeController.get_unit",
-      parameters: [
-        %Reference{"$ref": "#/components/parameters/arke_id"},
-        %Reference{"$ref": "#/components/parameters/unit_id"},
-        %Reference{"$ref": "#/components/parameters/arke-project-key"},
-        %Reference{"$ref": "#/components/parameters/limit"},
-        %Reference{"$ref": "#/components/parameters/offset"},
-        %Reference{"$ref": "#/components/parameters/order"},
-        %Reference{"$ref": "#/components/parameters/filter"}
-      ],
-      security: [%{"authorization" => []}],
-      responses: Responses.get_responses([201, 204])
-    }
-  end
-
-  def create_operation() do
-    %Operation{
-      tags: ["Unit"],
-      summary: "Create Unit",
-      description: "Create a new unit for the given Arke",
-      operationId: "ArkeServer.ArkeController.create",
-      parameters: [%Reference{"$ref": "#/components/parameters/arke-project-key"}],
-      requestBody:
-        OpenApiSpex.Operation.request_body(
-          "Parameters to create a unit",
-          "application/json",
-          ArkeServer.Schemas.CreateUnitExample,
-          required: true
-        ),
-      security: [%{"authorization" => []}],
-      responses: Responses.get_responses([200, 204])
-    }
-  end
-
-  def update_operation() do
-    %Operation{
-      tags: ["Unit"],
-      summary: "Login",
-      description: "Provide credentials to login to the app",
-      operationId: "ArkeServer.ArkeController.update",
-      parameters: [%Reference{"$ref": "#/components/parameters/arke-project-key"}],
-      requestBody:
-        OpenApiSpex.Operation.request_body(
-          "Parameters to login",
-          "application/json",
-          ArkeServer.Schemas.UserParams,
-          required: true
-        ),
-      security: [%{"authorization" => []}],
-      responses: Responses.get_responses([201, 204])
-    }
-  end
-
-  def delete_operation() do
-    %Operation{
-      tags: ["Unit"],
-      summary: "Delete Unit",
-      description: "Delete a specific Unit",
-      operationId: "ArkeServer.ArkeController.delete",
-      parameters: [
-        %Reference{"$ref": "#/components/parameters/arke_id"},
-        %Reference{"$ref": "#/components/parameters/unit_id"},
-        %Reference{"$ref": "#/components/parameters/arke-project-key"}
-      ],
-      security: [%{"authorization" => []}],
-      responses: Responses.get_responses([200, 201])
-    }
-  end
-
-  def get_all_unit_operation() do
-    %Operation{
-      tags: ["Unit"],
-      summary: "Get all",
-      description: "Get all the unit for the given Arke",
-      operationId: "ArkeServer.ArkeController.get_all_unit",
-      parameters: [
-        %Reference{"$ref": "#/components/parameters/arke_id"},
-        %Reference{"$ref": "#/components/parameters/arke-project-key"},
-        %Reference{"$ref": "#/components/parameters/limit"},
-        %Reference{"$ref": "#/components/parameters/offset"},
-        %Reference{"$ref": "#/components/parameters/order"},
-        %Reference{"$ref": "#/components/parameters/filter"}
-      ],
-      security: [%{"authorization" => []}],
-      responses: Responses.get_responses([201, 204])
-    }
-  end
-
-  def get_groups_operation() do
-    %Operation{
-      tags: ["Group"],
-      summary: "Get arke in group",
-      description: "Get all available groups",
-      operationId: "ArkeServer.ArkeController.get_groups",
-      parameters: [
-        %Reference{"$ref": "#/components/parameters/arke_id"},
-        %Reference{"$ref": "#/components/parameters/arke-project-key"},
-        %Reference{"$ref": "#/components/parameters/limit"},
-        %Reference{"$ref": "#/components/parameters/offset"},
-        %Reference{"$ref": "#/components/parameters/order"},
-        %Reference{"$ref": "#/components/parameters/filter"}
-      ],
-      security: [%{"authorization" => []}],
-      responses: Responses.get_responses([201, 204])
-    }
-  end
-
-  # ------- end OPENAPI spec -------
 
   def data_as_klist(data) do
     Enum.map(data, fn {key, value} -> {String.to_existing_atom(key), value} end)
   end
 
-  ## Once the project header has been set to mandatory retrieve it as follows:
-  #  project = conn.assigns[:arke_project]
-
   @doc """
        It returns a unit
-       """ && false
+       """ 
   def get_unit(conn, %{"unit_id" => _unit_id}) do
     # TODO handle query parameter with plugs
     load_links = Map.get(conn.query_params, "load_links", "false") == "true"
@@ -177,7 +59,7 @@ defmodule ArkeServer.ArkeController do
 
   @doc """
        Create a new unit
-       """ && false
+       """ 
   def create(%Plug.Conn{body_params: params} = conn, %{"arke_id" => id}) do
     # all arkes struct and gen server are on :arke_system so it won't be changed to project
     project = conn.assigns[:arke_project]
@@ -210,7 +92,7 @@ defmodule ArkeServer.ArkeController do
   # delete
   @doc """
        Delete a unit
-       """ && false
+       """ 
   def delete(conn, %{"unit_id" => _unit_id, "arke_id" => _arke_id}) do
     project = conn.assigns[:arke_project]
 
@@ -223,7 +105,7 @@ defmodule ArkeServer.ArkeController do
 
   @doc """
        Get units
-       """ && false
+       """ 
   def get_all_unit(conn, %{"arke_id" => id}) do
     project = conn.assigns[:arke_project]
 
@@ -283,7 +165,7 @@ defmodule ArkeServer.ArkeController do
   defp parse_coordinate(c) when is_binary(c), do: String.to_float(c)
   defp parse_coordinate(c), do: c
 
-  def calculate_coordinates(latitude, longitude, radius) do
+  defp calculate_coordinates(latitude, longitude, radius) do
 
     latitude_to_km = 110.574
     longitude_to_km = 111.32
@@ -302,7 +184,7 @@ defmodule ArkeServer.ArkeController do
 
   @doc """
        Call Arke function
-       """ && false
+       """ 
   def call_arke_function(conn, %{"arke_id" => arke_id, "function_name" => function_name}) do
     project = conn.assigns[:arke_project]
 
@@ -326,7 +208,7 @@ defmodule ArkeServer.ArkeController do
 
   @doc """
        Call Unit function
-       """ && false
+       """ 
   def call_unit_function(conn, %{
         "arke_id" => arke_id,
         "unit_id" => unit_id,
@@ -366,7 +248,7 @@ defmodule ArkeServer.ArkeController do
 
   @doc """
        Get Arke groups
-       """ && false
+       """ 
   def get_groups(conn, %{"arke_id" => id}) do
     project = conn.assigns[:arke_project]
     offset = Map.get(conn.query_params, "offset", nil)
