+++ lib/arke_server/controllers/auth_controller.ex
@@ -15,9 +15,13 @@
 defmodule ArkeServer.AuthController do
   @moduledoc """
              Documentation for `ArkeServer.AuthController`. Used from the controller and via API not from CLI
-             """ && false
+             """
 
   use ArkeServer, :controller
+
+  # Openapi request definition
+  use ArkeServer.Openapi.Spec, module: ArkeServer.Openapi.AuthControllerSpec
+
   alias ArkeServer.ResponseManager
   alias ArkeAuth.Core.{User, Auth, Otp}
   alias Arke.Boundary.ArkeManager
@@ -30,104 +34,6 @@ defmodule ArkeServer.AuthController do
 
   alias OpenApiSpex.{Operation, Reference}
   @mailer_module Application.get_env(:arke_server, :mailer_module)
-
-  # ------- start OPENAPI spec -------
-  def open_api_operation(action) do
-    operation = String.to_existing_atom("#{action}_operation")
-    apply(__MODULE__, operation, [])
-  end
-
-  def signin_operation() do
-    %Operation{
-      tags: ["Auth"],
-      summary: "Login",
-      description: "Provide credentials to login to the app",
-      operationId: "ArkeServer.AuthController.signin",
-      requestBody:
-        OpenApiSpex.Operation.request_body(
-          "Parameters to login",
-          "application/json",
-          ArkeServer.Schemas.UserParams,
-          required: true
-        ),
-      responses: Responses.get_responses([201, 204])
-    }
-  end
-
-  def signup_operation() do
-    %Operation{
-      tags: ["Auth"],
-      summary: "Signup",
-      description: "Create a new user for the app",
-      operationId: "ArkeServer.AuthController.signup",
-      requestBody:
-        OpenApiSpex.Operation.request_body(
-          "Parameters for the signup",
-          "application/json",
-          ArkeServer.Schemas.UserExample,
-          required: true
-        ),
-      responses: Responses.get_responses([201, 204])
-    }
-  end
-
-  def refresh_operation() do
-    %Operation{
-      tags: ["Auth"],
-      summary: "Refresh token",
-      description:
-        "Exchange the refresh token with a new couple access_token and refresh_token. Send the refresh token in the `authorization` header",
-      operationId: "ArkeServer.AuthController.refresh",
-      security: [%{"authorization" => []}],
-      responses: Responses.get_responses([201, 204])
-    }
-  end
-
-  def verify_operation() do
-    %Operation{
-      tags: ["Auth"],
-      summary: "Verify token",
-      description: "Check if the current access_token is still valid otherwise try to refresh it",
-      operationId: "ArkeServer.AuthController.verify",
-      security: [%{"authorization" => []}],
-      responses: Responses.get_responses([201, 204])
-    }
-  end
-
-  def change_password_operation() do
-    %Operation{
-      tags: ["Auth"],
-      summary: "Change user password",
-      description: "Changhe user pasword",
-      operationId: "ArkeServer.AuthController.change_password",
-      security: [%{"authorization" => []}],
-      responses: Responses.get_responses([200, 400])
-    }
-  end
-
-  def recover_password_operation() do
-    %Operation{
-      tags: ["Auth"],
-      summary: "Send email to reset passsword",
-      description: "Send an email containing a link to reset the password",
-      operationId: "ArkeServer.AuthController.recover_password",
-      security: [],
-      responses: Responses.get_responses([200, 400])
-    }
-  end
-
-  def reset_password_operation() do
-    %Operation{
-      tags: ["Auth"],
-      summary: "Reset the user password",
-      description: "Reset the user password",
-      operationId: "ArkeServer.AuthController.reset_password",
-      security: [],
-      responses: Responses.get_responses([200, 400])
-    }
-  end
-
-  # ------- end OPENAPI spec -------
 
   defp data_as_klist(data) do
     Enum.map(data, fn {key, value} -> {String.to_existing_atom(key), value} end)
