+++ lib/arke_server/controllers/struct_controller.ex
@@ -15,8 +15,13 @@
 defmodule ArkeServer.StructController do
   @moduledoc """
               Documentation for`ArkeServer.StructController
-             """ && false
+             """
   use ArkeServer, :controller
+
+  # Openapi request definition
+  use ArkeServer.Openapi.Spec, module: ArkeServer.Openapi.StructControllerSpec
+
+
   alias Arke.{StructManager}
   alias Arke.Boundary.ArkeManager
   alias ArkeServer.ResponseManager
@@ -25,51 +30,9 @@ defmodule ArkeServer.StructController do
 
   alias OpenApiSpex.{Operation, Reference}
 
-  # ------- start OPENAPI spec -------
-
-  def open_api_operation(action) do
-    operation = String.to_existing_atom("#{action}_operation")
-    apply(__MODULE__, operation, [])
-  end
-
-  def get_unit_struct_operation() do
-    %Operation{
-      tags: ["Struct"],
-      summary: "Get unit struct",
-      description:
-        "Get all the parameter and their types associated to the given element. Useful to create/update",
-      operationId: "ArkeServer.TopologyController.get_unit_struct",
-      parameters: [
-        %Reference{"$ref": "#/components/parameters/arke_id"},
-        %Reference{"$ref": "#/components/parameters/unit_id"},
-        %Reference{"$ref": "#/components/parameters/arke-project-key"}
-      ],
-      security: [%{"authorization" => []}],
-      responses: Responses.get_responses([201, 204])
-    }
-  end
-
-  def get_arke_struct_operation() do
-    %Operation{
-      tags: ["Struct"],
-      summary: "Get arke struct",
-      description:
-        "Get all the parameter and their types associated to the given element. Useful to create/update",
-      operationId: "ArkeServer.TopologyController.get_arke_struct",
-      parameters: [
-        %Reference{"$ref": "#/components/parameters/arke_id"},
-        %Reference{"$ref": "#/components/parameters/arke-project-key"}
-      ],
-      security: [%{"authorization" => []}],
-      responses: Responses.get_responses([201, 204])
-    }
-  end
-
-  # ------- end OPENAPI spec -------
-
   @doc """
        Get a struct of a unit
-       """ && false
+       """
   def get_unit_struct(conn, %{"arke_id" => arke_id, "arke_unit_id" => _id}) do
     project = conn.assigns[:arke_project]
     arke = ArkeManager.get(String.to_existing_atom(arke_id), project)
@@ -81,7 +44,7 @@ defmodule ArkeServer.StructController do
 
   @doc """
        Get a struct of an Arke
-       """ && false
+       """
   def get_arke_struct(conn, %{"arke_id" => id}) do
     project = conn.assigns[:arke_project]
