+++ lib/arke_server/controllers/topology_controller.ex
@@ -1,9 +1,13 @@
 defmodule ArkeServer.TopologyController do
   @moduledoc """
              Documentation for  `ArkeServer.TopologyController`.
-             """ && false
+             """
 
   use ArkeServer, :controller
+
+  # Openapi request definition
+  use ArkeServer.Openapi.Spec, module: ArkeServer.Openapi.TopologyControllerSpec
+
 
   alias Arke.{QueryManager, LinkManager, StructManager}
   alias Arke.Utils.ErrorGenerator, as: Error
@@ -14,128 +18,9 @@ defmodule ArkeServer.TopologyController do
 
   alias OpenApiSpex.{Operation, Reference}
 
-  # ------- start OPENAPI spec -------
-
-  def open_api_operation(action) do
-    operation = String.to_existing_atom("#{action}_operation")
-    apply(__MODULE__, operation, [])
-  end
-
-  def get_node_operation() do
-    %Operation{
-      tags: ["Topology"],
-      summary: "Get node",
-      description: "Get all elements (limited by depth parameter) linked to the given Unit",
-      operationId: "ArkeServer.TopologyController.get_node",
-      parameters: [
-        %Reference{"$ref": "#/components/parameters/arke_id"},
-        Operation.parameter(:arke_unit_id, :path, :string, "Parent Unit ID", required: true),
-        %Reference{"$ref": "#/components/parameters/link_id"},
-        Operation.parameter(:direction, :path, :string, "Direction where to get the node",
-          example: "child",
-          required: true
-        ),
-        %Reference{"$ref": "#/components/parameters/arke-project-key"}
-      ],
-      security: [%{"authorization" => []}],
-      responses: Responses.get_responses(201)
-    }
-  end
-
-  def create_node_operation() do
-    %Operation{
-      tags: ["Topology"],
-      summary: "Create connection",
-      description: "Create a link between two units",
-      operationId: "ArkeServer.TopologyController.create_node",
-      parameters: [
-        %Reference{"$ref": "#/components/parameters/arke_id"},
-        Operation.parameter(:arke_unit_id, :path, :string, "Parent Unit ID", required: true),
-        %Reference{"$ref": "#/components/parameters/link_id"},
-        Operation.parameter(:arke_id_two, :path, :string, "Child Arke ID", required: true),
-        Operation.parameter(:unit_id_two, :path, :string, "Child Arke ID", required: true),
-        %Reference{"$ref": "#/components/parameters/arke-project-key"}
-      ],
-      security: [%{"authorization" => []}],
-      responses: Responses.get_responses()
-    }
-  end
-
-  def update_node_operation() do
-    %Operation{
-      tags: ["Topology"],
-      summary: "Update connection",
-      description: "Update link metadata between two units",
-      operationId: "ArkeServer.TopologyController.update_node(",
-      parameters: [
-        %Reference{"$ref": "#/components/parameters/arke_id"},
-        Operation.parameter(:arke_unit_id, :path, :string, "Parent Unit ID", required: true),
-        %Reference{"$ref": "#/components/parameters/link_id"},
-        Operation.parameter(:arke_id_two, :path, :string, "Child Arke ID", required: true),
-        Operation.parameter(:unit_id_two, :path, :string, "Child Arke ID", required: true),
-        %Reference{"$ref": "#/components/parameters/arke-project-key"}
-      ],
-      security: [%{"authorization" => []}],
-      responses: Responses.get_responses([200])
-    }
-  end
-
-  def delete_node_operation() do
-    %Operation{
-      tags: ["Topology"],
-      summary: "Delete connection",
-      description: "Delete connection between two nodes",
-      operationId: "ArkeServer.TopologyController.delete_node",
-      parameters: [
-        Operation.parameter(:arke_id, :path, :string, "Parent Arke ID", required: true),
-        Operation.parameter(:arke_unit_id, :path, :string, "Parent Unit ID", required: true),
-        %Reference{"$ref": "#/components/parameters/link_id"},
-        Operation.parameter(:arke_id_two, :path, :string, "Child Arke ID", required: true),
-        Operation.parameter(:unit_id_two, :path, :string, "Child Unit ID", required: true),
-        %Reference{"$ref": "#/components/parameters/arke-project-key"}
-      ],
-      security: [%{"authorization" => []}],
-      responses: Responses.get_responses([200, 201])
-    }
-  end
-
-  def add_parameter_operation() do
-    %Operation{
-      tags: ["Parameter"],
-      summary: "Add parameter",
-      description: "Add parameter to the given Arke",
-      operationId: "ArkeServer.TopologyController.add_parameter",
-      parameters: [
-        %Reference{"$ref": "#/components/parameters/arke_id"},
-        %Reference{"$ref": "#/components/parameters/arke_parameter_id"},
-        %Reference{"$ref": "#/components/parameters/arke-project-key"}
-      ],
-      security: [%{"authorization" => []}],
-      responses: Responses.get_responses([200, 201])
-    }
-  end
-
-  def update_parameter_operation() do
-    %Operation{
-      tags: ["Parameter"],
-      summary: "Update associated parameter",
-      description: "Updates associated parameter of the given Arke",
-      operationId: "ArkeServer.TopologyController.update_parameter",
-      parameters: [
-        %Reference{"$ref": "#/components/parameters/arke_id"},
-        %Reference{"$ref": "#/components/parameters/arke_parameter_id"},
-        %Reference{"$ref": "#/components/parameters/arke-project-key"}
-      ],
-      security: [%{"authorization" => []}],
-      responses: Responses.get_responses([200])
-    }
-  end
-
-  # ------- end OPENAPI spec -------
-
   @doc """
             Get the unit linked to an Arke
-       """ && false
+       """
   def get_node(conn, %{"arke_id" => _arke_id, "arke_unit_id" => _id, "direction" => direction}) do
     project = conn.assigns[:arke_project]
     direction = String.to_existing_atom(direction)
@@ -169,7 +54,7 @@ defmodule ArkeServer.TopologyController do
 
   @doc """
        Link two unit together
-       """ && false
+       """
   def create_node(%Plug.Conn{body_params: params} = conn, %{
         "arke_id" => arke_id,
         "arke_id_two" => arke_id_two,
@@ -201,7 +86,7 @@ defmodule ArkeServer.TopologyController do
 
   @doc """
        Update metadata of an existing link
-       """ && false
+       """
 
   def update_node(%Plug.Conn{body_params: params} = conn, %{
         "arke_unit_id" => parent_id,
@@ -220,7 +105,7 @@ defmodule ArkeServer.TopologyController do
 
   @doc """
        Delete a connection between two units
-       """ && false
+       """
   def delete_node(%Plug.Conn{body_params: params} = conn, %{
         "arke_id" => _arke_id,
         "arke_id_two" => _arke_id_two,
@@ -243,7 +128,7 @@ defmodule ArkeServer.TopologyController do
 
   @doc """
        Associate a parameter to an Arke
-       """ && false
+       """
   def add_parameter(%Plug.Conn{body_params: params} = conn, %{
         "arke_parameter_id" => parameter_id,
         "arke_id" => arke_id
@@ -275,7 +160,7 @@ defmodule ArkeServer.TopologyController do
 
   @doc """
        Update an associated parameter of an Arke
-       """ && false
+       """
   def update_parameter(%Plug.Conn{body_params: params} = conn, %{
         "arke_parameter_id" => parameter_id,
         "arke_id" => arke_id
