+++ lib/arke_server/openapi/struct_controller_spec.ex
@@ -1,0 +1,56 @@
+# Copyright 2023 Arkemis S.r.l.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+defmodule ArkeServer.Openapi.StructControllerSpec do
+  @moduledoc """
+  Definition of the ApiSpec for `ArkeServer.StructController`.
+  """
+
+  alias ArkeServer.Openapi.Responses
+  alias OpenApiSpex.{Operation, Reference}
+
+  def get_unit_struct_operation() do
+    %Operation{
+      tags: ["Struct"],
+      summary: "Get unit struct",
+      description:
+        "Get all the parameter and their types associated to the given element. Useful to create/update",
+      operationId: "ArkeServer.TopologyController.get_unit_struct",
+      parameters: [
+        %Reference{"$ref": "#/components/parameters/arke_id"},
+        %Reference{"$ref": "#/components/parameters/unit_id"},
+        %Reference{"$ref": "#/components/parameters/arke-project-key"}
+      ],
+      security: [%{"authorization" => []}],
+      responses: Responses.get_responses([201, 204])
+    }
+  end
+
+  def get_arke_struct_operation() do
+    %Operation{
+      tags: ["Struct"],
+      summary: "Get arke struct",
+      description:
+        "Get all the parameter and their types associated to the given element. Useful to create/update",
+      operationId: "ArkeServer.TopologyController.get_arke_struct",
+      parameters: [
+        %Reference{"$ref": "#/components/parameters/arke_id"},
+        %Reference{"$ref": "#/components/parameters/arke-project-key"}
+      ],
+      security: [%{"authorization" => []}],
+      responses: Responses.get_responses([201, 204])
+    }
+  end
+
+end
