+++ lib/arke_server/plugs/get_project.ex
@@ -15,7 +15,7 @@
 defmodule ArkeServer.Plugs.GetProject do
   @moduledoc """
              Plug to get the project from the request header
-             """ && false
+             """
   import Plug.Conn
   alias Arke.Utils.ErrorGenerator, as: Error
